requirements.txt
python-fitbit>=0.1